requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
aiohttp==3.7.4.post0
python-dotenv==0.19.0